+++ src/main/java/org/opentripplanner/api/model/RouterInfo.java
@@ -49,7 +49,7 @@ public class RouterInfo {
     public boolean health;
 
     @Getter
-    private final Map<SharedVehiclesSnapshotLabel, Long> acceptedSharedVehiclesSnapshotLabels;
+    private final Map<SharedVehiclesSnapshotLabel, Integer> acceptedSharedVehiclesSnapshotLabels;
 
     public RouterInfo(String routerId, Graph graph) {
         this.acceptedSharedVehiclesSnapshotLabels = graph.getSupportedSnapshotLabels();
